improve how playlist items display
- remove useEffect Hook
- remove useState Hook
- move table types to its own file
- remove unused imports
-remove console log
